requirements.txt
flask==3.0.0
flask-cors==4.0.0
llama-cpp-python==0.2.78
torch==2.1.0
transformers==4.35.0
numpy==1.24.3
sentencepiece==0.1.99
accelerate==0.24.1
datasets==2.14.6
peft==0.6.2
bitsandbytes==0.41.1
rouge-score==0.1.2
nltk==3.8.1
pyyaml==6.0.1
flask-limiter==3.5.0
marshmallow==3.20.1
python-dotenv==1.0.0
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
